pyproject.toml
[project]
name = "kvm-light-manager"
version = "2.0.0"
description = "Enterprise Async VM orchestration service with KVM/libvirt simulation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "sqlalchemy==2.0.28",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "redis==5.0.3",
    "pyzmq==25.1.2",
    "pydantic-settings==2.2.1",
    "python-json-logger==2.0.7",
    "prometheus-client==0.19.0",
    "tenacity==8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.2",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "testcontainers==4.3.3",
    "httpx==0.27.0",
    "websockets==12.0",
    "fakeredis==2.21.1",
    "pre-commit==3.6.0",
    "ruff==0.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov=agent --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]